% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcastr.R
\name{etl_extract.etl_statcastr}
\alias{etl_extract.etl_statcastr}
\title{Scrape Statcast data}
\usage{
\method{etl_extract}{etl_statcastr}(obj, years = 2017, months = 6, ...)
}
\arguments{
\item{obj}{ETL object}

\item{years}{A numeric vector of years}

\item{months}{A numeric vector of months}

\item{...}{arguments passed to \code{\link[baseballr]{scrape_statcast_savant_batter_all}}}
}
\description{
Scrape Statcast data
}
\examples{

sc <- etl("statcastr")
\dontrun{
sc \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load(tablenames = "statcast")
}
}
