% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_statcast}
\alias{fetch_statcast}
\title{Fetch MLB statcast data.}
\usage{
fetch_statcast(dates, verbose = TRUE)
}
\arguments{
\item{dates}{The dates from which to retrieve statcast data}

\item{verbose}{TRUE or FALSE: Whether or not to report messages.}
}
\value{
A statcast data.table
}
\description{
This function queries Baseball Savant for statcast data on dates specified by \code{dates} and returns an R object with the results.
}
\details{
The \code{dates} argument is intended to be as flexible as possible. A single date of the form yyyy-mm-dd may be used: \code{fetch_statcast("2017-03-20")}. To retrieve statcast data for an entire season, specify the year: \code{fetch_statcast(2017)}. To select a range of dates, use the separator " to " as follows: \code{fetch_statcast("2017 April 12 to 2017 April 20")}. To select specific dates, combine them in a character vector: \code{fetch_statcast(c('2017 Apr 12','2018 April 12'))}. All of these modes can be used in conjunction, as long as year is listed first (then month, then day) within each individual date.
}
\examples{
fetch_statcast("2017 April 12")
fetch_statcast("2019 Sept 4 to 2020")
}
