% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{load_statcast}
\alias{load_statcast}
\title{Load a statcast data object to disk}
\usage{
load_statcast(input)
}
\arguments{
\item{input}{PATH to .csv or .RDS file generated by \code{save_statcast()}}
}
\value{
A statcast object
}
\description{
This function load a statcast data object (\code{sc}) from the specified location (\code{input})
}
\details{
This function will load statcast data from disk (which was previously stored using \code{save_statcast()}). To get new statcast data, use \code{fetch_statcast()}
}
\examples{
sc <- load_statcast('sc_file.csv')
}
\seealso{
[save_statcast()], [fetch_statcast()]
}
