% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{save_statcast}
\alias{save_statcast}
\title{Save a statcast data object to disk}
\usage{
save_statcast(sc, output = NA, verbose = TRUE)
}
\arguments{
\item{sc}{A statcast data object}

\item{output}{A character string specifying PATH to which the object will be saved. The file extension determines how the file will be saved. Options include \code{.csv} for comma-separated values, and \code{.RDS} for RDS format. If output == NA, the file name will be automatically generated.}

\item{verbose}{TRUE or FALSE: Whether or not to report messages.}
}
\value{
(nothing)
}
\description{
This function saves a statcast data object (\code{sc}) to a specified location (\code{output})
}
\details{
This function will write statcast data to disk, which can then be re-loaded into R with \code{load_statcast()}
}
\examples{
sc <- statcast()
save_statcast(sc,'empty')
}
\seealso{
[load_statcast()]
}
