% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{sort_statcast}
\alias{sort_statcast}
\title{Sort statcast data.}
\usage{
sort_statcast(x)
}
\arguments{
\item{sc}{A statcast data object}
}
\value{
A sorted statcast data object
}
\description{
Sort statcast data, first by date, then at_bat_number, and finally by pitch_number
}
\details{
This function ensures that pitch data are sorted in chronological order.
}
\examples{
sort_statcast()
}
