% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{update_statcast}
\alias{update_statcast}
\title{Update a statcast data object with additional and/or new data}
\usage{
update_statcast(sc, dates, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{sc}{A statcast data object}

\item{dates}{A flexible parameter for specifying dates (see [fetch_statcast()] for more details)}

\item{overwrite}{TRUE or FALSE: whether or not dates in \code{sc} will be overwritten with new data corresponding to \code{dates} argument. If TRUE the server will be queried with all dates listed in \code{dates}, even if some/all already exist in \code{sc}. If FALSE, only \code{dates} that don't already exist in \code{sc} will be queried.}

\item{verbose}{TRUE or FALSE: Whether or not to report messages.}
}
\value{
A statcast object, with any new dates worth of data incorporated.
}
\description{
This function adds new statcast data to an existing statcast object.
}
\details{
This function is useful for updating an existing statcast data object when new data becomes available.
}
\examples{
sc <- statcast()
update_statcast(sc,'2017 April 12')
}
\seealso{
[fetch_statcast()]
}
